requirements.txt
gunicorn
Flask
line-bot-sdk
requests